Send steps from child component
